피드백: 범용 함수 분리 (price)
